requirements.txt
streamlit
ultralytics
pillow
pandas
scikit-learn
openai
langchain
